pyproject.toml
[project]
name = "docprep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
